restructure generics
