+++ mythx_cli/fuzz/ide/__init__.py
@@ -1,0 +1,2 @@
+from mythx_cli.fuzz.ide.generic import IDEArtifacts
+from mythx_cli.fuzz.ide.brownie import BrownieJob
